Add test for psudo_labels
